refactor: extract IssueExportPresenters.swift from IssueExportController.swift
Extracts IssueExportStatusPresenter enum (17 lines) + IssueExportPresentationController
@MainActor class (48 lines) from IssueExportController.swift into a new file.

Byte-identical move; SHA verified against origin/main:IssueExportController.swift
lines 4-70.

Applied memory rule (annotation preservation): initially attempted range 4-72
which included the @MainActor annotation for IssueExportController class (at
line 72, its class declaration at line 73). Corrected to 4-70 which stops at
the closing brace of IssueExportPresentationController, preserving the
@MainActor for IssueExportController in the remaining file.

Reduces IssueExportController.swift from 537 to 468 lines.

Tests: 667 (unchanged).
